requirements.txt
analytix==3.5.0
df2gspread==1.0.4
google_api_python_client>=1.6.7
google_auth_oauthlib==0.5.2
gspread==5.4.0
gspread_dataframe==3.3.0
mysql_connector_repackaged==0.3.1
numpy==1.22.4
oauth2client==4.1.3
pandas==1.4.2
protobuf==4.21.2
pygsheets==2.0.5
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
schedule==1.1.0
selenium==4.3.0
SQLAlchemy==1.3.24
tqdm==4.64.0
pytube==12.1.0